Cleaned up common functions
